requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-Session==0.5.0
email_validator==2.1.0

# Database
SQLAlchemy>=2.0.35
psycopg2-binary>=2.9.9

# OpenAI
openai==1.3.5
requests==2.31.0

# Web forms
WTForms==3.1.1

# Utilities
python-dotenv==1.0.0
Pillow>=11.0.0
python-dateutil==2.8.2

# For production
gunicorn==21.2.0

# Browser QA (optional): pip install then python -m playwright install chromium
playwright>=1.49.0,<2
